pyproject.toml
[tool.poetry]
name = "ai-groundwork"
version = "0.1.0"
description = ""
authors = ["dhkoo <amainlog@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.12"
python-dotenv = "^1.0.0"
openai = "^1.3.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
